research-agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
tavily-python==0.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
